betting-bot/requirements.txt
# Core Dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
aiomysql>=0.1.0

# API and Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0

# Image Processing
Pillow>=10.0.0

# Database and Caching
redis>=5.0.0
aioredis>=2.0.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Enhanced Player Props Dependencies
rapidfuzz>=3.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
psutil>=5.9.0
structlog>=23.0.0

# Web Framework (if using web interface)
flask>=2.3.0
flask-cors>=4.0.0

# Background Tasks
celery>=5.3.0

# Monitoring and Logging
prometheus-client>=0.17.0

# Date and Time
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration
pyyaml>=6.0.0
toml>=0.10.0

# Utilities
click>=8.1.0
tqdm>=4.65.0 